style: move breadcrumb styles to module
