+++ src/pages/ReportPage/Breadcrumbs/Breadcrumbs.module.css
@@ -1,0 +1,7 @@
+.breadcrumbs__breadcrumb-button {
+  cursor: pointer;
+}
+
+.breadcrumbs__breadcrumb-separator {
+  color: white;
+}
